Move cursor & selection structs to text_body mod.
